sghss/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.16
python-dotenv>=1.0
